webserver/pyproject.toml
[tool.poetry]
name = "webserver"
version = "0.1.0"
description = "Lambda functions to ingest and enrich data"
authors = ["Josh Phillips <Josh.L.Phillips@me.com>"]

[tool.poetry.dependencies]
python = "3.12.7"
python-dotenv = "^0.19.1"
requests = "2.31.0"
psycopg2-binary = "2.9.9"
boto3 = "1.28.0"
pandas = "2.1.1"
scipy = "1.11.3"
SQLAlchemy = "1.4.54"
numpy = "1.26.0"
scikit-learn = "1.5.2"
openlineage-python = "1.18.0"
openai = "1.39.0"
flask = "2.2.5"
flask-sqlalchemy = "2.5.1"
flask-migrate = "^3.1.0"
WTForms = "3.0.1"
Flask-WTF = "1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"